+++ packages/libs/lib-server/src/system/settings/service/sys-settings-service.ts
@@ -169,11 +169,10 @@ export class SysSettingsService extends BaseService<SysSettingsEntity> {
     };
     setGlobalProxy(opts);
     setGlobalHeaders(this.parseKeyValueText(privateSetting.commonHeaders));
-
+    
     if (privateSetting.dnsResultOrder) {
       dns.setDefaultResultOrder(privateSetting.dnsResultOrder as any);
     }
-
     if (privateSetting.pipelineMaxRunningCount) {
       executorQueue.setMaxRunningCount(privateSetting.pipelineMaxRunningCount);
     }