+++ packages/ui/certd-server/src/configuration.ts
@@ -31,7 +31,6 @@ process.on("uncaughtException", error => {
     logger.error("您的服务器不支持监听IPV6格式的地址（::），请配置环境变量: certd_koa_hostname=0.0.0.0");
   }
 });
-
 // function startHeapLog() {
 //   function format(bytes: any) {
 //     return (bytes / 1024 / 1024).toFixed(2) + ' MB';
@@ -132,9 +131,7 @@ export class MainConfiguration {
     setLogger((text: string) => {
       logger.info(text);
     });
-
     logger.info("当前环境：", this.app.getEnv()); // prod
-
     
   }
 }