+++ packages/ui/certd-server/src/plugins/plugin-aliyun/plugin/deploy-to-cdn/index.ts
@@ -209,7 +209,7 @@ export class DeployCertToAliyunCDN extends AbstractTaskPlugin {
   }
 
   async getClient(access: AliyunAccess) {
-    const client = new AliyunClient({ logger: this.logger });
+    const client = new AliyunClient({ logger: this.logger, importRuntime: access.importRuntime.bind(access) });
     await client.init({
       accessKeyId: access.accessKeyId,
       accessKeySecret: access.accessKeySecret,
@@ -254,12 +254,15 @@ export class DeployCertToAliyunCDN extends AbstractTaskPlugin {
     const client = await this.getClient(access);
 
     const pager = new Pager(data);
-    const params = {
-      DomainName: data.searchKey,
+    const params: any = {
       PageSize: pager.pageSize || 100,
       PageNumber: pager.pageNo || 1,
       DomainSearchType: "fuzzy_match",
     };
+
+    if (data.searchKey) {
+      params.DomainName = data.searchKey;
+    }
 
     const requestOption = {
       method: "POST",