+++ packages/ui/certd-server/src/plugins/plugin-cert/access/acme-account-access.ts
@@ -100,7 +100,7 @@ export class AcmeAccountAccess extends BaseAccess {
       "\nGoogle:请查看[google获取eab帮助文档](https://certd.docmirror.cn/guide/use/google/),用过一次后会绑定邮箱，后续复用EAB要用同一个邮箱" +
       "\nSSL.com:[SSL.com账号页面](https://secure.ssl.com/account),然后点击api credentials链接，然后点击编辑按钮，查看Secret key和HMAC key" +
       "\nlitessl:[litesslEAB页面](https://freessl.cn/automation/eab-manager),然后点击新增EAB",
-    required: false,
+    required: true,
     encrypt: true,
     mergeScript: `
     return {
@@ -121,7 +121,7 @@ export class AcmeAccountAccess extends BaseAccess {
     component: {
       placeholder: "需要EAB的颁发机构生成账号时填写",
     },
-    required: false,
+    required: true,
     encrypt: true,
     mergeScript: `
     return {
@@ -138,18 +138,18 @@ export class AcmeAccountAccess extends BaseAccess {
   eabHmacKey = "";
 
   @AccessInput({
-    title: "ACME账号信息",
+    title: "生成ACME账号",
     component: {
       name: "refresh-input",
       action: "GenerateAccount",
       buttonText: "生成ACME账号",
       successMessage: "ACME账号已生成，请保存授权配置",
-      type:"textarea",
-      rows:4,
+      type: "textarea",
+      rows: 4,
     },
-    col:{span:24},
+    col: { span: 24 },
     required: true,
-    helper: "请生成ACME账号，账号一旦生成不允许修改",
+    helper: "请点击右边按钮生成ACME账号，账号一旦生成不允许修改",
     encrypt: true,
     mergeScript: `
     return {
@@ -170,7 +170,6 @@ export class AcmeAccountAccess extends BaseAccess {
     required: false,
     helper: "是否开启修改ACME账号，注意，开启后，会影响DNS持久验证记录",
     encrypt: false,
-   
   })
   editAccount = false;